+++ Firmware/lib/oled_energy_display/oled_library.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "globals.h"
 #include "oled_energy_display.h"
 #include "oled_touch_wake.h"
 
@@ -18,7 +19,7 @@ bool begin(const Settings& settings);
 void update();
 
 bool startBackgroundUpdater(uint32_t intervalMs = 20,
-                           uint32_t stackSizeWords = 1424,
+                           uint32_t stackSizeWords = OLED_UPDATE_TASK_STACK_SIZE,
                            uint32_t priority = 1,
                            int8_t coreId = 1);
 void stopBackgroundUpdater();